feat: weekly view for check-in screen (issue #72)
Add Weekly tab to the check-in screen with a 7-column scrollable grid
showing Mon–Sun habit completion. Split CheckInScreen into DailyView
and WeeklyView components. Fix timezone bug in getCheckInsForWeek.
